# [Core][MultiModalHasher] Hash images without converting image mode