## Purpose

This PR hashes mode, palette and data of images separately which prevents the need for converting all images to RGBA. See https://github.com/vllm-project/vllm/pull/24925#discussion_r2350397896

## Test Plan

Correctness should be covered by the existing hasher tests on CI.

The performance can be measured using:

```python
import numpy as np
from PIL import Image
from vllm.multimodal.hasher import MultiModalHasher

np.random.seed(42)
data = np.random.randint(0, 255, size=(3840, 2160, 3), dtype=np.uint8)
data = Image.fromarray(data)

%timeit MultiModalHasher.hash_kwargs(data=data)
```

## Test Result

For a 4k PIL image this speeds up hashing by ~35%. This is not massive, but might add up in cases with lots of multimodal input.

```shell
# main
25.1 ms ± 124 μs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# This PR
16.3 ms ± 74.6 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```